# swagger.eko.yml
openapi: 3.0.3
info:
  title: EKO Chatbot API
  description: Complete API documentation for EKO Chatbot application
  version: 1.0.0
  contact:
    name: EKO API Support
    email: support@eko.com
servers:
  - url: https://api.eko.com/v1
    description: Production server
  - url: https://staging-api.eko.com/v1
    description: Staging server
security:
  - BearerAuth: []
paths:
...
  /chat/create:
    post:
      tags:
        - Chat
      summary: Create New Chat
      description: Create a new chat conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - short_description
                - is_temporary
              properties:
                title:
                  type: string
                  example: New Programming Discussion
                short_description:
                  type: string
                  example: Help with JavaScript concepts
                is_temporary:
                  type: boolean
                  example: false
      responses:
        '201':
          description: Chat created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateChatResponse'
  /chat/{chat_id}/messages:
    get:
      tags:
        - Chat
      summary: Get Conversation Messages
      description: Retrieve paginated messages from a chat conversation
      parameters:
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
            example: chat_789
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            example: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            example: 20
      responses:
        '200':
          description: Conversation retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
  /chat/{chat_id}/message:
    post:
      tags:
        - Chat
      summary: Send Message to Chatbot
      description: Send a message to the chatbot in a specific chat
      parameters:
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
            example: chat_789
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - chat_id
                - message
              properties:
                chat_id:
                  type: string
                  example: chat_789
                message:
                  type: string
                  example: Can you explain closures in JavaScript?
                pictures:
                  type: array
                  items:
                    type: string
                  example: ["image_url_1", "image_url_2"]
                voices:
                  type: array
                  items:
                    type: string
                  example: ["voice_url_1"]
      responses:
        '200':
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMessageResponse'
  /chat/{chat_id}:
    delete:
      tags:
        - Chat
      summary: Delete Chat
      description: Delete a specific chat conversation
      parameters:
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
            example: chat_789
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - chat_id
              properties:
                chat_id:
                  type: string
                  example: chat_789
      responses:
        '200':
          description: Chat deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteChatResponse'
  /chat/all:
    delete:
      tags:
        - Chat
      summary: Delete All Chats
      description: Delete all user chats
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_auth_token
              properties:
                user_auth_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: All chats deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAllChatsResponse'
  /message/{message_id}:
    put:
      tags:
        - Messages
      summary: Update Message
      description: Update a specific message
      parameters:
        - name: message_id
          in: path
          required: true
          schema:
            type: string
            example: msg_003
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message_id
                - message
              properties:
                message_id:
                  type: string
                  example: msg_003
                message:
                  type: string
                  example: Updated message content
                pictures:
                  type: array
                  items:
                    type: string
                  example: ["new_image_url_1"]
                voices:
                  type: array
                  items:
                    type: string
                  example: ["new_voice_url_1"]
      responses:
        '200':
          description: Message updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateMessageResponse'
    delete:
      tags:
        - Messages
      summary: Delete Message
      description: Delete a specific message
      parameters:
        - name: message_id
          in: path
          required: true
          schema:
            type: string
            example: msg_003
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message_id
              properties:
                message_id:
                  type: string
                  example: msg_003
      responses:
        '200':
          description: Message deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteMessageResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateChatResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Chat created successfully
        data:
          type: object
          properties:
            chat_id:
              type: string
              example: chat_789
            title:
              type: string
              example: New Programming Discussion
            short_description:
              type: string
              example: Help with JavaScript concepts
            is_temporary:
              type: boolean
              example: false
            created_at:
              type: string
              format: date-time
              example: 2025-07-31T10:00:00Z
    ConversationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Conversation retrieved successfully
        data:
          type: object
          properties:
            messages:
              type: array
              items:
                type: object
                properties:
                  message_id:
                    type: string
                    example: msg_001
                  chat_id:
                    type: string
                    example: chat_789
                  sender:
                    type: string
                    enum: [user, bot]
                    example: user
                  message:
                    type: string
                    example: Hello, can you help me with JavaScript?
                  pictures:
                    type: array
                    items:
                      type: string
                    example: []
                  voices:
                    type: array
                    items:
                      type: string
                    example: []
                  timestamp:
                    type: string
                    format: date-time
                    example: 2025-07-31T10:01:00Z
            pagination:
              type: object
              properties:
                current_page:
                  type: integer
                  example: 1
                total_pages:
                  type: integer
                  example: 5
                total_messages:
                  type: integer
                  example: 100
                has_next:
                  type: boolean
                  example: true
    SendMessageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Message sent successfully
        data:
          type: object
          properties:
            message_id:
              type: string
              example: msg_003
            chat_id:
              type: string
              example: chat_789
            sender:
              type: string
              example: user
            message:
              type: string
              example: Can you explain closures in JavaScript?
            pictures:
              type: array
              items:
                type: string
              example: ["image_url_1", "image_url_2"]
            voices:
              type: array
              items:
                type: string
              example: ["voice_url_1"]
            timestamp:
              type: string
              format: date-time
              example: 2025-07-31T10:05:00Z
            bot_response:
              type: object
              properties:
                message_id:
                  type: string
                  example: msg_004
                chat_id:
                  type: string
                  example: chat_789
                sender:
                  type: string
                  example: bot
                message:
                  type: string
                  example: A closure in JavaScript is a function that has access to variables in its outer scope even after the outer function has returned...
                pictures:
                  type: array
                  items:
                    type: string
                  example: []
                voices:
                  type: array
                  items:
                    type: string
                  example: []
                timestamp:
                  type: string
                  format: date-time
                  example: 2025-07-31T10:05:15Z
    DeleteChatResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Chat deleted successfully
        data:
          type: object
          properties:
            deleted_chat_id:
              type: string
              example: chat_789
    DeleteAllChatsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: All chats deleted successfully
        data:
          type: object
          properties:
            deleted_chats_count:
              type: integer
              example: 15
    UpdateMessageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Message updated successfully
        data:
          type: object
          properties:
            message_id:
              type: string
              example: msg_003
            updated_message:
              type: string
              example: Updated message content
            pictures:
              type: array
              items:
                type: string
              example: ["new_image_url_1"]
            voices:
              type: array
              items:
                type: string
              example: ["new_voice_url_1"]
            updated_at:
              type: string
              format: date-time
              example: 2025-07-31T10:10:00Z
    DeleteMessageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Message deleted successfully
        data:
          type: object
          properties:
            deleted_message_id:
              type: string
              example: msg_003
...